chore: update dependencies in go.mod and go.sum, enhance CRD definitions in seaweed_webhook.go, and refactor admin service and stateful set management in controller_admin.go
